Added run_simulation method to Orchestrator class
This was actually kind of exciting to see everything come together and run properly. I had to throw in a couple of additional tweaks to handle some edge cases where one dataset had station or trip information that the other didn't. But otherwise it was all pretty straightforward, and we're now getting actual "No bike was available" messages being emitted.

I need to validate that they're actually accurate, and to build in some additional utilities to monitor and visualize the bike traffic tick-over-tick, but that's all for a later commit.